+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 Upcoming v0.1.4: CreationResponse#success?, #blank?
 - CreationResponse: body is parsed from JSON
 - Copy/paste #blank? from Rails/ActiveSupport
+- FIX Uploader.from_file_path factory method (named args)
 
 2020-07-22 v0.1.3: CreationRequest accepts :body, :extra_headers
 - i.e. Vimeo tus server requires info to be included in the request.